pyproject.toml
[tool.poetry]
name = "chatgpt-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Levon Gegamyan <levongeg@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chatgpt_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
openai = "^0.27.0"
aiogram = "3.0.0b7"
setuptools = "^67.7.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
